Add tooltip properties to mark
